feat(codegen): introduce module func for built-in functions
